Fix error message displayed in getErrorMessage method in OlympicService, and rename model Olympic to OlympicCountry for consistency across the project
